Rest

+++ Rest API/controllers/feed.js
@@ -3,41 +3,33 @@ const Post = require("../model/post");
 const fs = require("fs");
 const path = require("path");
 const { response } = require("express");
+const User = require("../model/user");
+const { getIO } = require("../socket");
 
 const POSTS_PER_PAGE = 2;
 
-exports.getFeed = (req, res, next) => {
+exports.getFeed = async (req, res, next) => {
   const page = req.query.page || 1;
-  Post.find()
-  .countDocuments()
-  .then((total) => {
-    console.log(req.query);
-    Post.find()
-        .skip(POSTS_PER_PAGE * (page - 1))
-        .limit(POSTS_PER_PAGE)
-        .then((posts) => {
-          // console.log(posts);
-          return res.status(200).json({
-            posts: posts,
-            totalItems: total
-          });
-        })
-        .catch((err) => {
-          if (!err.statusCode) {
-            err.statusCode = 500;
-          }
-          next(err);
-        });
-    })
-    .catch((err) => {
-      if (!err.statusCode) {
-        err.statusCode = 500;
-      }
-      next(err);
+  try {
+    const total = await Post.find().countDocuments();
+    const posts = await Post.find()
+      .populate("creator")
+      .sort({createdAt: -1})
+      .skip(POSTS_PER_PAGE * (page - 1))
+      .limit(POSTS_PER_PAGE);
+    res.status(200).json({
+      posts: posts,
+      totalItems: total,
     });
+  } catch (err) {
+    if (!err.statusCode) {
+      err.statusCode = 500;
+    }
+    next(err);
+  }
 };
 
-exports.postPost = (req, res, next) => {
+exports.postPost = async (req, res, next) => {
   //create post
   if (!req.file) {
     const error = new Error("No image found");
@@ -48,6 +40,8 @@ exports.postPost = (req, res, next) => {
   const content = req.body.content;
   const imageUrl = req.file.path.replace("\\", "/");
   const errors = validationResult(req);
+  let creator;
+  let newPost;
   if (!errors.isEmpty()) {
     const error = new Error("Validation Failed");
     error.statusCode = 422;
@@ -57,26 +51,32 @@ exports.postPost = (req, res, next) => {
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: "Fakeha",
-    },
+    creator: req.userId,
   });
-  post
-    .save()
-    .then((result) => {
-      return res
-        .status(201) //resource successfully created
-        .json({
-          message: "Created Post!",
-          post: result,
-        });
-    })
-    .catch((err) => {
-      if (!err.statusCode) {
-        err.statusCode = 500;
-      }
-      next(err);
+  try {
+    const result = await post.save();
+    newPost = result;
+    const user = await User.findById(req.userId);
+    creator = user;
+    user.posts.push(newPost);
+    await user.save();
+    getIO().emit("posts", {
+      action: "create",
+      post: { ...newPost._doc, creator: { _id: req.userId, name: user.name } },
     });
+    res
+      .status(201) //resource successfully created
+      .json({
+        message: "Created Post!",
+        post: newPost,
+        creator: { _id: creator._id, name: creator.name },
+      });
+  } catch (err) {
+    if (!err.statusCode) {
+      err.statusCode = 500;
+    }
+    next(err);
+  }
 };
 
 exports.getPost = (req, res, next) => {
@@ -120,11 +120,16 @@ exports.putPost = (req, res, next) => {
     throw error;
   }
 
-  Post.findById(postId)
+  Post.findById(postId).populate("creator")
     .then((post) => {
       if (!post) {
         const error = new Error("No Post Found!");
         error.statusCode = 404;
+        throw error;
+      }
+      if (req.userId !== post.creator._id.toString()) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
         throw error;
       }
       if (imageUrl !== post.imageUrl) {
@@ -136,6 +141,7 @@ exports.putPost = (req, res, next) => {
       return post.save();
     })
     .then((post) => {
+      getIO().emit("posts", {action: "update", post: post})
       return res.status(200).json({
         message: "Post Updated successfully!",
         post: post,
@@ -158,11 +164,24 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (req.userId !== post.creator.toString()) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       const imgUrl = post.imageUrl;
       clearImage(imgUrl);
       return Post.findByIdAndDelete(id);
     })
     .then((response) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(id);
+      return user.save();
+    })
+    .then((response) => {
+      getIO().emit("posts", {action: "delete", post: id})
       res.status(200).json({ message: "Deleted Post" });
     })
     .catch((err) => {
@@ -176,4 +195,39 @@ exports.deletePost = (req, res, next) => {
 const clearImage = (filePath) => {
   const p = path.join(__dirname, "..", filePath);
   fs.unlink(p, (err) => console.log(err));
+};
+
+exports.getStatus = (req, res, next) => {
+  User.findById(req.userId)
+    .then((user) => {
+      console.log();
+      res.status(200).json({
+        status: user.status,
+      });
+    })
+    .catch((err) => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    });
+};
+
+exports.putStatus = (req, res, next) => {
+  User.findById(req.userId)
+    .then((user) => {
+      console.log(user);
+      const status = req.body.status;
+      user.status = status;
+      return user.save();
+    })
+    .then(() => {
+      res.status(200).json({ message: "Status updated!" });
+    })
+    .catch((err) => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    });
 };
